refactor : write 페이지 수정 (# 96)
